Rework test suite a bit
