+++ sparse_strips/vello_cpu/snapshots/gradient_linear_2_stops.png
