+++ sparse_strips/vello_cpu/snapshots/gradient_linear_4_stops.png
