+++ sparse_strips/vello_cpu/snapshots/gradient_linear_complex_shape.png
