+++ sparse_strips/vello_cpu/snapshots/gradient_linear_shape.png
@@ -1,0 +1,3 @@
+version https://git-lfs.github.com/spec/v1
+oid sha256:80a4f4a9a3b4287abbda922be42b35c06fe6e58c86b4854e4f462bca6ebac40a
+size 1734
