+++ sparse_strips/vello_cpu/snapshots/gradient_linear_spread_method_pad.png
