+++ sparse_strips/vello_cpu/tests/sweep_gradient.rs
@@ -1,4 +1,4 @@
-use crate::util::{check_ref, get_ctx, star_path};
+use crate::util::{check_ref, get_ctx, star_path, stops_green_blue};
 use std::f64::consts::PI;
 use vello_common::color::palette::css::{BLACK, BLUE, GREEN, RED, WHITE, YELLOW};
 use vello_common::kurbo::{Affine, Point, Rect};
@@ -23,38 +23,4 @@ fn gradient_sweep_2_stops() {
     ctx.fill_rect(&rect);
 
     check_ref(&ctx, "gradient_sweep_2_stops");
-}
-
-fn stops_green_blue() -> Vec<Stop> {
-    vec![
-        Stop {
-            offset: 0.0,
-            color: GREEN,
-        },
-        Stop {
-            offset: 1.0,
-            color: BLUE,
-        },
-    ]
-}
-
-fn stops_blue_green_red_yellow() -> Vec<Stop> {
-    vec![
-        Stop {
-            offset: 0.0,
-            color: BLUE,
-        },
-        Stop {
-            offset: 0.33,
-            color: GREEN,
-        },
-        Stop {
-            offset: 0.66,
-            color: RED,
-        },
-        Stop {
-            offset: 1.0,
-            color: YELLOW,
-        },
-    ]
 }
