+++ sparse_strips/vello_cpu/tests/util.rs
@@ -16,16 +16,18 @@ use std::io::Cursor;
 use std::path::PathBuf;
 use std::sync::LazyLock;
 use vello_common::color::palette;
+use vello_common::color::palette::css::{BLUE, GREEN, RED, YELLOW};
 use vello_common::kurbo::{BezPath, Rect, Shape};
 use vello_common::pixmap::Pixmap;
 use vello_cpu::RenderContext;
+use vello_cpu::paint::Stop;
 
 static REFS_PATH: LazyLock<PathBuf> =
     LazyLock::new(|| PathBuf::from(env!("CARGO_MANIFEST_DIR")).join("snapshots"));
 static DIFFS_PATH: LazyLock<PathBuf> =
     LazyLock::new(|| PathBuf::from(env!("CARGO_MANIFEST_DIR")).join("diffs"));
 
-pub(crate) fn get_ctx(width: u16, height: u16, transparent: bool) -> RenderContext {
+pub fn get_ctx(width: u16, height: u16, transparent: bool) -> RenderContext {
     let mut ctx = RenderContext::new(width, height);
     if !transparent {
         let path = Rect::new(0.0, 0.0, width as f64, height as f64).to_path(0.1);
@@ -37,14 +39,61 @@ pub(crate) fn get_ctx(width: u16, height: u16, transparent: bool) -> RenderConte
     ctx
 }
 
-pub(crate) fn render_pixmap(ctx: &RenderContext) -> Pixmap {
+pub fn render_pixmap(ctx: &RenderContext) -> Pixmap {
     let mut pixmap = Pixmap::new(ctx.width(), ctx.height());
     ctx.render_to_pixmap(&mut pixmap);
 
     pixmap
 }
 
-pub(crate) fn check_ref(ctx: &RenderContext, name: &str) {
+pub fn stops_green_blue() -> Vec<Stop> {
+    vec![
+        Stop {
+            offset: 0.0,
+            color: GREEN,
+        },
+        Stop {
+            offset: 1.0,
+            color: BLUE,
+        },
+    ]
+}
+
+pub fn stops_green_blue_with_alpha() -> Vec<Stop> {
+    vec![
+        Stop {
+            offset: 0.0,
+            color: GREEN.with_alpha(0.25),
+        },
+        Stop {
+            offset: 1.0,
+            color: BLUE.with_alpha(0.75),
+        },
+    ]
+}
+
+pub fn stops_blue_green_red_yellow() -> Vec<Stop> {
+    vec![
+        Stop {
+            offset: 0.0,
+            color: BLUE,
+        },
+        Stop {
+            offset: 0.33,
+            color: GREEN,
+        },
+        Stop {
+            offset: 0.66,
+            color: RED,
+        },
+        Stop {
+            offset: 1.0,
+            color: YELLOW,
+        },
+    ]
+}
+
+pub fn check_ref(ctx: &RenderContext, name: &str) {
     let mut pixmap = render_pixmap(ctx);
     pixmap.unpremultiply();
 
@@ -103,7 +152,7 @@ pub(crate) fn check_ref(ctx: &RenderContext, name: &str) {
     }
 }
 
-pub(crate) fn star_path() -> BezPath {
+pub fn star_path() -> BezPath {
     let mut path = BezPath::new();
     path.move_to((50.0, 10.0));
     path.line_to((75.0, 90.0));
